+++ app/client/firebase/fetchIsLoggedIn.ts
