+++ app/client/microcms/post/fetchPostDetail.ts
@@ -1,3 +1,3 @@
-import { GET_CLIENT } from '../api/detail/[contentId]/routeClient'
+import { GET_CLIENT } from '../../../api/detail/[contentId]/routeClient'
 
 export const fetchPostDetail = async (contentId: string) => await GET_CLIENT(contentId)
