+++ hooks/useCreatePost.ts
@@ -1,4 +1,4 @@
 import { useMutation } from '@tanstack/react-query'
-import { createPost } from '../app/client/createPost'
+import { createPost } from '../app/client/microcms/post/createPost'
 
 export const useCreatePost = () => useMutation((question: string) => createPost(question))
