+++ hooks/useCreateReply.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query'
-import { createReply } from '../app/client/createReply'
-import { patchReplyId } from '../app/client/patchReplyId'
+import { createReply } from '../app/client/microcms/reply/createReply'
+import { patchReplyId } from '../app/client/microcms/post/patchReplyId'
 
 export const useCreateReply = () =>
   useMutation(async ({ reply, replyFor }: { reply: string; replyFor: string }) => {
