+++ hooks/useDeleteReply.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query'
-import { deleteReply } from '../app/client/deleteReply'
+import { deleteReply } from '../app/client/microcms/reply/deleteReply'
 
 export const useDeleteReply = () =>
   useMutation(async (contentId: string) => await deleteReply(contentId))
