+++ hooks/useMoveToAnswer.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query'
-import { deleteReply } from '../app/client/deleteReply'
-import { createPost } from '../app/client/createPost'
+import { deleteReply } from '../app/client/microcms/reply/deleteReply'
+import { createPost } from '../app/client/microcms/post/createPost'
 
 export const useMoveToAnswer = () =>
   useMutation(async ({ contentId, question }: { contentId: string; question: string }) => {
